Fix failure on bare repositories
Bare repositories are often named "foo.git". If we have a bare repository
in "~/src/foo.git", then scanning while "~/src", `d` would be "foo.git",
causing `dirproject` to be set to "~/src/fo".

Since `dirnames` contains a list of the directory names we can directly
check for ".git" being in the list. This avoids the bug and is a bit
faster (no need to check for other elements in `dirnames` once ".git" has
been found)