Add option to let user specify the repos that will be printed.
With the 'onlychanged' flag, the user can specify that he wants
to be printed only the repositories that has any changes.

So this way, a user with a long list of repositories can be more
effective to see what he needs to do next, since he doesn't needs
to seek the changed repos on the printed list.