Make sure that the terminal colors are intact.
We're not resetting the terminal colors after printing the output,
so all the "original" colors of the terminal is messed after running
gitcheck.

Printing the reset escape sequence after printing the output let
the colors as it was before. Even if no colors were output the reset
escape code is harmless.